move server impl to separate module, set proper initial_call
